# backend/src/main/java/com/multiagent/codeassistant/dto/AiResponse.java
package com.multiagent.codeassistant.dto;

public class AiResponse {
    private String response;

    public AiResponse() {
    }

    public AiResponse(String response) {
        this.response = response;
    }

    public String getResponse() {
        return response;
    }

    public void setResponse(String response) {
        this.response = response;
    }
}